school_management/requirements.txt
alembic==1.14.0
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.6
MarkupSafe==2.1.5
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing-extensions==4.12.2
werkzeug==3.0.6
wtforms==3.1.2
zipp==3.20.2
